% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Get countries}
\usage{
get_countries(access_token, name = "")
}
\arguments{
\item{name}{Pattern to filter country names}

\item{access}{token API key to authenticate to AviSample rest API}
}
\value{
tibble
}
\description{
Returns a list of countries that have at least one sample description
}
