% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_descriptions.R
\name{get_sample_descriptions}
\alias{get_sample_descriptions}
\title{Get Sample descriptions}
\usage{
get_sample_descriptions(
  taxon_common = "",
  taxon_scientific = "",
  country = "",
  tissue = "",
  perPage = 50,
  limit = 1000
)
}
\arguments{
\item{taxon_common}{Pattern to filter sample descriptions by common names}

\item{taxon_scientific}{Pattern to filter sample descriptions by scientific names}

\item{country}{Pattern to filter sample descriptions by country}

\item{tissue}{Pattern to filter sample descriptions by tissue type}

\item{api_key}{API key to authenticate to AviSample rest API}
}
\value{
A tibble
}
\description{
Retrieves the list of sample descriptions. The function is intentionally slow to prevent server overload. By default it is limited to 1000 items but this parameter can be set.
}
