% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{Get Taxa}
\usage{
get_taxa(scientific = "", common = "")
}
\arguments{
\item{scientific}{A string pattern to filter scientific names}

\item{common}{A string pattern to filter common names}
}
\description{
Internal function that retrieves list of all taxa matching a filter string
}
