% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_types.R
\name{get_tissue_types}
\alias{get_tissue_types}
\title{Get Tissue Types}
\usage{
get_tissue_types(name = NULL)
}
\arguments{
\item{name}{A string pattern to filter tissue type names}
}
\description{
Internal function that retrieves list of all taxa matching a filter string
}
